requirements.txt
altgraph==0.17.3
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
flit_core==3.9.0
fqdn==1.5.1
glob2==0.7
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.5
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.2.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter-lsp==2.1.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.4.3
matplotlib-inline==0.1.6
mistune==2.0.5
mplcairo==0.5
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.22.3
overrides==7.3.1
packaging==23.1
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
progressbar==2.5
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycairo==1.23.0
pycparser==2.21
pyeit==1.2.4
Pygments==2.15.1
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyrsistent==0.19.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
sciopy==0.6.4
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==67.7.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.60.0
traitlets==5.9.0
trimesh==3.21.6
typing_extensions==4.6.2
typing-utils==0.1.0
uri-template==1.2.0
urllib3==2.0.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
wheel==0.38.4
widgetsnbextension==4.0.7
win-inet-pton==1.1.0
zipp==3.15.0
